requirements.txt
setuptools==75.8.0
aiogram==3.16.0
aiohttp==3.11.11
aiohappyeyeballs==2.4.4
python-dotenv==1.0.1
aiosqlite==0.20.0
matplotlib==3.10.0
pyowm==3.3.0
